Created basic chrome extension. Closes #1
This is just a basic template that covers the necessary pieces for
the chrome extension to work. Shows an alert warning a user about
visiting heroku.com (and only heroku.com). Real logic for deciding if a
website should fall into the warning category will be done in another
module