Added first tests and basic functionality to match. Addresses #4
This introduces tests to check if urls without http at the start return
false when checked if they are part of the black list. This actually
feels unwanted since it anyway wouldn't match the blacklist. But keeping
it for now. Also wrote the matching function to take in a parameter of a
url and a dictionary of urls to check against. The matcher function will
be responsible for stripping URLs down to root domain names to check if
they can be found inside the dictionary. Currently it returns true or
false, but I'll want it to return something more informative later like
an empty string for no matches and the 'key' value for the dictionary if
it does find a matching url.