Made url matching better! Addresses #4
The url matcher can now take a url of the formats

* https://domain.com
* http://domain.com
* http://www.domain.com
* http://www.sub.domain.com
* http://sub.domain.com
* http://sub.domain.com/

and a few more and match it to the key

'domain.com'

It does this by first removing the protocol part of the url.

It then takes out the www. part if it exists

It proceeds to remove the ending slash

It then checks if it can find the url thus far inside the dictionary.
This is done so that we can match a sub domain only instead of always
matching the full root domain. Example would be a company blog that
resides on a blogging service. The blogging service shouldn't be
penalised, but <domain>.blogginservice.com should.