Arranged files to be able to package chrome extension
First of all, all thanks go to icon monstr for providing the icons to
get this started. Need to update README to reflect this. Please check
out iconmonstr at iconmonstr.com

This preps the directories for building the chrome extension for the
first time. Need to add to readme that till the grunt file is ready, it
is necessary to copy the directory.js and matcher.js files to the chrome
directory before packaging it.

Updated directory.js file to actually contain the dictionary of
blacklisted sites

Updated manifest.json to include matcher.js and directory.js in. Need to
test if I can simply load it in from outside of the directory so I can
then eliminate the copying step